requirements.txt
blinker==1.9.0
click==8.3.0
dnspython==2.8.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.15.1
python-dotenv==1.1.1
Werkzeug==3.1.3
flask-cors==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.7